requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
bleach==6.1.0
boto3==1.34.155
botocore==1.34.155
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
distro==1.9.0
Django==4.2.15
django-admin-autocomplete-filter==0.7.1
django-cors-headers==4.3.1
django-filter==24.3
django-rest-swagger==2.2.0
django-storages==1.14.2
django-summernote==0.8.20.0
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openai==1.40.8
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
tqdm==4.66.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
webencodings==0.5.1
